requirement.txt
FastAPI
Uvicorn
Pydantic
google-cloud-aiplatform
python-dotenv